+++ lib/src/parsers/raster/thumbhash.dart
@@ -1,2 +1,2 @@
-export 'thumbhash/thumbhash_decoder_source.dart';
+export 'thumbhash/thumbhash_decoder.dart';
 export 'thumbhash/thumbhash_encoder.dart';